requirements.txt
streamlit
python-dotenv
google-generativeai
moviepy
SpeechRecognition
pydub
openai-whisper